+++ frontend/shelter-frontend/src/app/app-routing.module.ts
@@ -1,9 +1,9 @@
 import {NgModule} from '@angular/core';
 import {RouterModule, Routes} from '@angular/router';
-import {LoginComponent} from "./login/login.component";
-import {RegisterComponent} from "./register/register.component";
+import {LoginComponent} from "./component/login/login.component";
+import {RegisterComponent} from "./component/register/register.component";
 import {AppComponent} from "./app.component";
-import {PetsComponent} from "./pets/pets.component";
+import {PetsComponent} from "./pages/petsPage/pets.component";
 
 const routes: Routes = [
   {
