+++ frontend/shelter-frontend/src/app/component/pet-card/pet-card.component.ts
