+++ frontend/shelter-frontend/src/app/component/top-bar/top-bar.component.ts
@@ -1,5 +1,5 @@
 import {Component, OnInit} from '@angular/core';
-import {AccountService} from "../_services/account.service";
+import {AccountService} from "../../_services/account.service";
 
 @Component({
   selector: 'app-top-bar',
