+++ frontend/shelter-frontend/src/app/pages/petsPage/pets.component.css
